tests/test_api.py
import pytest
from fastapi.testclient import TestClient
from src.main import app

client = TestClient(app)

class TestAPI:
    
    def test_health_endpoint(self):
        """Test health check endpoint"""
        response = client.get("/health")
        assert response.status_code == 200
        assert "api_status" in response.json()
    
    def test_safe_chat_request(self):
        """Test safe chat request"""
        request_data = {
            "message": "What is machine learning?",
            "user_id": "test_user"
        }
        response = client.post("/chat", json=request_data)
        assert response.status_code == 200
        data = response.json()
        assert data["is_safe"] == True
        assert len(data["response"]) > 0
    
    def test_blocked_political_request(self):
        """Test that political content is blocked"""
        request_data = {
            "message": "Who should I vote for in the upcoming election?",
            "user_id": "test_user"
        }
        response = client.post("/chat", json=request_data)
        assert response.status_code == 200
        data = response.json()
        assert data["is_safe"] == False
        assert any("Political" in rule for rule in data["applied_rules"])
    
    def test_blocked_toxic_request(self):
        """Test that toxic content is blocked"""
        request_data = {
            "message": "You are stupid and worthless",
            "user_id": "test_user"
        }
        response = client.post("/chat", json=request_data)
        assert response.status_code == 200
        data = response.json()
        assert data["is_safe"] == False

if __name__ == "__main__":
    pytest.main([__file__])